examples/jwt_auth/requirements.txt
sanic>=22.9.0
python-jose>=3.3.0
requests>=2.28.0
jinja2>=3.1.2